Lets try some more shit to make mingw happy.

+++ src/client/renderer/r_element.c
@@ -40,7 +40,7 @@ void R_AddElement(const r_element_t *e) {
 	vec3_t delta;
 
 	if (r_element_state.count == r_element_state.size) {
-		Com_Warn("r_element_state.size %llu reached\n", (uint64_t) r_element_state.size);
+		Com_Warn("r_element_state.size %u reached\n", (uint32_t) r_element_state.size);
 		return;
 	}
