+++ src/tests/check_r_media.c
@@ -77,7 +77,7 @@ START_TEST(check_R_RegisterMedia)
 		R_BeginLoading();
 		R_FreeMedia();
 
-		ck_assert_msg(Z_Size() == 0, "Not all memory freed: %llu", (uint64_t) Z_Size());
+		ck_assert_msg(Z_Size() == 0, "Not all memory freed: %u", (uint32_t) Z_Size());
 
 	}END_TEST
